fixing crud movie with role

+++ src/controllers/movie.js
@@ -43,43 +43,55 @@ movies.getMovie = async (req, res) => {
 
 // Create a Movie
 movies.createData = async (req, res) => {
-    try {
-        const { title, description, release_date, directed_by, duration, casts, images, category, price } = req.body
-        const data = await models.addData({ title, description, release_date, directed_by, duration, casts, images, category, price })
-        return response(res, 201, data)
-    } catch (error) {
-        return response(res, 500, error)
+    if (req.user.role === 'admin') {
+        try {
+            const { title, description, release_date, directed_by, duration, casts, images, category, price } = req.body
+            const data = await models.addData({ title, description, release_date, directed_by, duration, casts, images, category, price })
+            return response(res, 201, data)
+        } catch (error) {
+            return response(res, 500, error)
+        }
+    } else {
+        return response(res, 403, 'Maaf anda bukan admin')
     }
 }
 
 // Update a Movie
 movies.updateData = async (req, res) => {
-    try {
-        const { id } = req.params
-        const { title, description, release_date, directed_by, duration, casts, images, price, category } = req.body
-        const data = await models.updateData({ id, title, description, release_date, directed_by, duration, casts, images, price, category })
-        if (!data.length) {
-            return response(res, 404, 'Data tidak ditemukan')
-        } else {
-            return response(res, 200, data)
+    if (req.user.role === 'admin') {
+        try {
+            const { id } = req.params
+            const { title, description, release_date, directed_by, duration, casts, images, price, category } = req.body
+            const data = await models.updateData({ id, title, description, release_date, directed_by, duration, casts, images, price, category })
+            if (!data.length) {
+                return response(res, 404, 'Data tidak ditemukan')
+            } else {
+                return response(res, 200, data)
+            }
+        } catch (error) {
+            return response(res, 500, error)
         }
-    } catch (error) {
-        return response(res, 500, error)
+    } else {
+        return response(res, 403, 'Maaf anda bukan admin')
     }
 }
 
 // Delete a Movie
 movies.deleteMovie = async (req, res) => {
-    try {
-        const { id } = req.params
-        const data = await models.deleteData({ id })
-        if (data.rowCount < 1) {
-            return response(res, 404, 'Data tidak ditemukan')
-        } else {
-            return response(res, 200, 'Data berhasil di Delete')
+    if (req.user.role === 'admin') {
+        try {
+            const { id } = req.params
+            const data = await models.deleteData({ id })
+            if (data.rowCount < 1) {
+                return response(res, 404, 'Data tidak ditemukan')
+            } else {
+                return response(res, 200, 'Data berhasil di Delete')
+            }
+        } catch (error) {
+            return response(res, 500, error)
         }
-    } catch (error) {
-        return response(res, 500, error)
+    } else {
+        return response(res, 403, 'Maaf anda bukan admin')
     }
 }
